chore: consolidate test and benchmark code
