+++ .gitignore
@@ -1,4 +1,4 @@
+.integrationtests
 *.out
 *.test
-/bin
-.integrationtests
+/out
