+++ websocket.go
@@ -355,3 +355,47 @@ func statusCodeForError(err error) (StatusCode, string) {
 	}
 	return code, reason
 }
+
+// Hooks define the callbacks that are called during the lifecycle of a
+// websocket connection.
+type Hooks struct {
+	// OnClose is called when the connection is closed.
+	OnClose func(ClientKey, StatusCode, error)
+	// OnReadError is called when a read error occurs.
+	OnReadError func(ClientKey, error)
+	// OnReadFrame is called when a frame is read.
+	OnReadFrame func(ClientKey, *Frame)
+	// OnReadMessage is called when a complete message is read.
+	OnReadMessage func(ClientKey, *Message)
+	// OnWriteError is called when a write error occurs.
+	OnWriteError func(ClientKey, error)
+	// OnWriteFrame is called when a frame is written.
+	OnWriteFrame func(ClientKey, *Frame)
+	// OnWriteMessage is called when a complete message is written.
+	OnWriteMessage func(ClientKey, *Message)
+}
+
+// setupHooks ensures that all hooks have a default no-op function if unset.
+func setupHooks(hooks *Hooks) {
+	if hooks.OnClose == nil {
+		hooks.OnClose = func(ClientKey, StatusCode, error) {}
+	}
+	if hooks.OnReadError == nil {
+		hooks.OnReadError = func(ClientKey, error) {}
+	}
+	if hooks.OnReadFrame == nil {
+		hooks.OnReadFrame = func(ClientKey, *Frame) {}
+	}
+	if hooks.OnReadMessage == nil {
+		hooks.OnReadMessage = func(ClientKey, *Message) {}
+	}
+	if hooks.OnWriteError == nil {
+		hooks.OnWriteError = func(ClientKey, error) {}
+	}
+	if hooks.OnWriteFrame == nil {
+		hooks.OnWriteFrame = func(ClientKey, *Frame) {}
+	}
+	if hooks.OnWriteMessage == nil {
+		hooks.OnWriteMessage = func(ClientKey, *Message) {}
+	}
+}
